Posts * refactoring errors * create constants.ts
