+++ src/app/app.module.ts
@@ -26,6 +26,10 @@ import { ForumModule } from "./components/forum/forum.module";
 import { RegistrationFormModule } from "./components/registration-form/registration-form.module";
 import { TranslatorModule } from "./components/translator/translator.module";
 import { PostsModule } from "./components/posts-user/posts.module";
+import { StoreModule } from "@ngrx/store";
+import { appReducers } from "./components/store";
+import { StoreDevtoolsModule } from "@ngrx/store-devtools";
+
 @NgModule({
   declarations: [
     AppComponent,
@@ -52,6 +56,8 @@ import { PostsModule } from "./components/posts-user/posts.module";
     RegistrationFormModule,
     TranslatorModule,
     PostsModule,
+    StoreModule.forRoot(appReducers),
+    StoreDevtoolsModule.instrument(),
   ],
   bootstrap: [AppComponent],
 })
