+++ src/app/components/posts-user/components/posts/posts.component.html
@@ -1,11 +1,7 @@
 <ng-container *ngIf="postsList$ | async">
   <mat-list *ngFor="let post of postsList$ | async">
 
-    <span
-      [routerLink]="['/posts-user', post.id]">
-
-        {{ post.title }}
-    </span>
+    <span [routerLink]="['/posts-user', post.id]">{{ post.title }}</span>
     <mat-divider></mat-divider>
 
   </mat-list>
