+++ src/app/components/posts-user/constants/constants.ts
@@ -1,0 +1,5 @@
+export interface Post {
+  id: number;
+  title: string;
+  body: string;
+}
